requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
anthropic>=0.40
google-genai>=1.0
scikit-learn>=1.3
folium>=0.15
streamlit>=1.30
